Rewrite the sending algorithm of the packet out test
The sending algorithm for the packet out test should perform
faster and more accurately send the required rate.

This also makes the backlog check on the control channel optional
so the test can either back off when a backlog is detected or keep
writing (which is buffered inside libevent). As part of this extra
statistics are printed to indicate the number of packets actually
sent compared with those lost.